Create and update forms

+++ src/app/components/admin/plant/plant.component.ts
@@ -25,14 +25,17 @@ export class PlantComponent implements OnInit {
   }
 
   async deletePlant(plantId) {
-    const deleted = await this.plantsService.deletePlant(plantId);
-    let message;
-    if (deleted) {
-      message = 'La planta ha sido eliminada';
-    } else {
-      message = 'Ha ocurrido un error, por favor intentelo más tarde';
+    const control = await this.uiService.showConfirmAlert('¿Estas seguro de realizar está acción?');
+    if (control) {
+      const deleted = await this.plantsService.deletePlant(plantId);
+      let message;
+      if (deleted) {
+        message = 'La planta ha sido eliminada';
+      } else {
+        message = 'Ha ocurrido un error, por favor intentelo más tarde';
+      }
+      this.uiService.showToast(message);
     }
-    this.uiService.showToast(message);
   }
 
 }
